Segment Query And Mutations Based On Data Objects
