+++ src/main/java/com/graphqljava/tutorial/bookDetails/resolver/mutation/AuthorMutation.java
@@ -1,0 +1,25 @@
+package com.graphqljava.tutorial.bookDetails.resolver.mutation;
+
+import com.coxautodev.graphql.tools.GraphQLMutationResolver;
+import com.graphqljava.tutorial.bookDetails.mongoDB.entity.Author;
+import com.graphqljava.tutorial.bookDetails.mongoDB.repository.AuthorRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthorMutation implements GraphQLMutationResolver {
+
+    @Autowired
+    private AuthorRepository authorRepository;
+
+    public Author createAuthor(String firstName, String lastName, int age) {
+        Author author = new Author();
+        author.setFirstName(firstName);
+        author.setLastName(lastName);
+        author.setAge(age);
+
+        authorRepository.save(author);
+
+        return author;
+    }
+}
