+++ src/main/java/com/graphqljava/tutorial/bookDetails/resolver/mutation/BookMutation.java
@@ -1,41 +1,21 @@
-package com.graphqljava.tutorial.bookDetails.resolver;
+package com.graphqljava.tutorial.bookDetails.resolver.mutation;
 
 
 import com.coxautodev.graphql.tools.GraphQLMutationResolver;
-import com.graphqljava.tutorial.bookDetails.mongoDB.entity.Author;
 import com.graphqljava.tutorial.bookDetails.mongoDB.entity.Book;
-import com.graphqljava.tutorial.bookDetails.mongoDB.repository.AuthorRepository;
 import com.graphqljava.tutorial.bookDetails.mongoDB.repository.BookRepository;
 import graphql.GraphQLException;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.io.FileNotFoundException;
 import java.util.Optional;
 
 @Component
-public class Mutation implements GraphQLMutationResolver {
-
-    private AuthorRepository authorRepository;
-    private BookRepository bookRepository;
+public class BookMutation implements GraphQLMutationResolver {
 
     @Autowired
-    public Mutation(AuthorRepository authorRepository, BookRepository bookRepository) {
-        this.authorRepository = authorRepository;
-        this.bookRepository = bookRepository;
-    }
-
-    public Author createAuthor(String firstName, String lastName, int age) {
-        Author author = new Author();
-        author.setFirstName(firstName);
-        author.setLastName(lastName);
-        author.setAge(age);
-
-        authorRepository.save(author);
-
-        return author;
-    }
+    private BookRepository bookRepository;
 
     public Book createBook(String title, String description, String authorId) {
         Book book = new Book();
